+++ CHANGELOG.md
@@ -1,4 +1,9 @@
 
+## 2.10.0 - August 26, 2018
+
+  - `:element_key_mod` and `:attr_key_mod` options were added to allow keys to
+    be modified when loading.
+
 ## 2.9.4 - July 16, 2018
 
   - Fixed issue with malformed object mode input.